cache-loader/requirements.txt
redis==5.0.3
snowflake-connector-python==3.7.0
python-dotenv==1.0.1
